Getting started on the test environment.
I think the ideal way to guarentee that the tests never run against a
production database is to create a separate database within the postgres
instance exclusively for tests. It will be seeded before tests are run
and torn down after they finish. A separate .env file will be used for
this image to specify the test database name.