Tests (#4)
* Getting started on the test environment.

I think the ideal way to guarentee that the tests never run against a
production database is to create a separate database within the postgres
instance exclusively for tests. It will be seeded before tests are run
and torn down after they finish. A separate .env file will be used for
this image to specify the test database name.

* use make commands in circleci config

* Add make as a ci dependency

* Fixed test env and cleaned up Makefile

* use vanilla `lein test` on the ci

* Create a CI test .env that doesn't include the volume mount

* Fix name error

* remove eftest again

* Fix weird if else syntax

* Not gonna work if it's not getting used

* migrations need to be run before the tests

* set CLJ_ENV to test

* Make docker-compose.test.yml the default compose file for the test
targets

* Run the up/down tasks in pre/post test hooks within a transaction block.
This allows tests to run in parellel.

* Fix makefile

* cleaned up the create-user test and added another assertion

* fix indentation

* remove redundant dependencies